refactor: correct import paths for SVG icons
