+++ src/common/MovieTile/index.js
@@ -1,4 +1,4 @@
-import { ReactComponent as PlaceholderIcon } from "../assets/placeholderMovieIcon.svg";
+import { ReactComponent as PlaceholderIcon } from "../../assets/placeholderMovieIcon.svg";
 
 import { ImagePlaceholder } from "../ImagePlaceholder/styled";
 import {
